Update Accpeted Order View Model, repo DAO
The Order repo was updated to refactor name of method, and make use of a
newer query, like in new order fragment. The return type of the function
was also changed

The viewModel was updated to incoorporate the above changes in the repo,
as well as make a new variable to observe errors

There was a typo in DAO, due th which Issue #20 was not resolved. This
uptade fixes that too

Fixes #17 #20