Add Room compatiblity for Pie and above
For android version pie and above, the deafault mode for any SQLite
database has write_before_logging mode set to true, that was causing
errors in writing data to room. So update builder to rectify this.

Also changed some security rules in the manifest, and the default
network type

Fixes #23