Fix issue 19
Changed the flow of the app a bit. Initially, the orders were not
displayed as the table containing the menu items, items id and the
status was blank. Since orders require that information, and the query
uses data from that table as a join, it returned an empty list until the
menu_table was populated. So, now I am fetching the meneus also as soon
as the app starts

Fixes #19